AWS Hack/requirements.txt
# Agentic Store Operations Platform
# Python 3.10+

# CrewAI multi-agent
crewai>=0.80.0
crewai-tools>=0.14.0

# LLM: AWS Bedrock (no OpenAI key needed)
langchain-aws>=0.2.0
langchain-core>=0.3.0

# AWS
boto3>=1.34.0

# API server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Config and utilities
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Optional: local dev / testing
pytest>=7.0.0
httpx>=0.26.0
